# Replace convoluted indexOfTheOnlyBit implementation with ctz# primop