Performance is the same or slightly better in microbenchmark. I
suspect it can perform better in the wild.

I was also interested in `revNat` but GHC doesn't expose an
appropriate primop, presumably because GCC doesn't. It could still be
faster on clang (`__builtin_bitreverse64` and friends).