@Fuuzetsu 2019-01-24
I realised that there's `countTrailingZeroes` function in Data.Bits.FiniteBits typeclass, it's probably better to use that instead of GHC.Prim.