@Fuuzetsu 2019-01-24
IIRC I ran the benchmarks at the time and they were pretty much the same (so that's a good thing). I don't have the numbers at hand, I can try to run them again after changing to countTrailingZeroes and post the results.

I see that CI failed due to old GHC versions which didn't have the primop: are these versions we care about and I should add CPP?

w.r.t. reversing bits, let's forget about that for now. I made a GHC ticket if you're interested: https://ghc.haskell.org/trac/ghc/ticket/16164