@treeowl 2019-01-24
Yes, I think we still have to care about GHC 7.8 for another major version. Let's use CPP for now.